+++ mini_project/src/MPPI.h
@@ -9,17 +9,17 @@
 
 // steering and gas pedal commands
 struct ControlInput {
-    float steering;
-    float throttle;
+    double steering;
+    double throttle;
 };
 
 struct MPPIParms{
     int samples;
     int steps;
-    float dt;
-    float lambda;
-    float std_steer;
-    float std_throttle;
+    double dt;
+    double lambda;
+    double std_steer;
+    double std_throttle;
 
 };
 
@@ -27,27 +27,27 @@ class MPPI {
 private:
     int num_samples;            // ghost cars (e.g., 100)
     int horizon;                // steps into the future
-    float dt;                   // The time step 
+    double dt;                   // The time step 
 
-    float lambda;               // "Temperature" - controls how aggressively we filter bad paths
-    float noise_std_steer;      // stadard deviation of steering angle
-    float noise_std_throttle;   // standard deviation of throttle
+    double lambda;               // "Temperature" - controls how aggressively we filter bad paths
+    double noise_std_steer;      // stadard deviation of steering angle
+    double noise_std_throttle;   // standard deviation of throttle
 
-    float target_speed;
+    double target_speed;
 
     std::vector<ControlInput> nominal_trajectory;
 
     // C++ random number generators for creating the "ghost" noise
     std::mt19937 rng; 
-    std::normal_distribution<float> steer_dist;
-    std::normal_distribution<float> throttle_dist;
+    std::normal_distribution<double> steer_dist;
+    std::normal_distribution<double> throttle_dist;
 
 public:
     // Constructor
     MPPI(const MPPIParms& params);
 
     // THE BRAIN: Takes the current state, simulates the ghosts, and returns the best action
-    ControlInput get_best_control(const CarState& current_state, Car& car_model, Track& track, const std::vector<std::vector<std::pair<float, float>>>& other_paths, int mycar_id);
-    std::vector<std::pair<float, float>> get_predicted_path(const CarState& current_state, Car& car_model);
-    void set_target_speed(const float speed);
+    ControlInput get_best_control(const CarState& current_state, Car& car_model, Track& track, const std::vector<std::vector<std::pair<double, double>>>& other_paths, int mycar_id);
+    std::vector<std::pair<double, double>> get_predicted_path(const CarState& current_state, Car& car_model);
+    void set_target_speed(const double speed);
 };
